move laith's types to another file
