+++ src/components/schedule/schedule.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from "react";
 import { PendingApptCard } from "../appointments/pendingCard";
 import { ConfirmedApptCard } from "../appointments/confirmedCard";
 import { interviewController } from "@/controllers/interview";
-import { Interview } from "./types";
+import { Interview } from "@/types/types";
 import { authController } from "@/controllers/auth";
 
 export function Schedule() {
